move shared to a package in /exchanges
