+++ exchanges/populate/tsconfig.json
@@ -3,11 +3,11 @@
   "include": ["src"],
   "compilerOptions": {
     "baseUrl": "./",
-    "rootDirs": ["./", "../common"],
     "paths": {
       "urql": ["../../node_modules/urql/src"],
       "*-urql": ["../../node_modules/*-urql/src"],
-      "@urql/*": ["../../node_modules/@urql/*/src"]
+      "@urql/*": ["../../node_modules/@urql/*/src"],
+      "shared": ["../../node_modules/shared/src"]
     }
   }
 }
