+++ exchanges/shared/package.json
@@ -1,0 +1,4 @@
+{
+  "private": true,
+  "name": "shared"
+}
