+++ exchanges/shared/tsconfig.json
@@ -6,8 +6,7 @@
     "paths": {
       "urql": ["../../node_modules/urql/src"],
       "*-urql": ["../../node_modules/*-urql/src"],
-      "@urql/*": ["../../node_modules/@urql/*/src"],
-      "shared": ["../../node_modules/shared/src"]
+      "@urql/*": ["../../node_modules/@urql/*/src"]
     }
   }
 }
