+++ scripts/jest/preset.js
@@ -10,6 +10,7 @@ module.exports = {
     "^urql$": "<rootDir>/../../node_modules/urql/src",
     "^(.*-urql)$": "<rootDir>/../../node_modules/$1/src",
     "^@urql/(.*)$": "<rootDir>/../../node_modules/@urql/$1/src",
+    "^shared$": "<rootDir>/../../node_modules/shared/src",
   },
   watchPlugins: ['jest-watch-yarn-workspaces'],
   testRegex: '(src/.*(\\.|/)(test|spec))\\.tsx?$',
